Fixed penetration. It should work better and lag less nao.
BEWARE! I will probably be switching SVN links to accommodate GDCW SWEPS soon! When I do, go to the GDC website (It's in the info file in the main addon folder) and get the new repository checkout link. THEN
Go to Tortoise SVN ---> Switch ---> (Copy and paste new link in the box)

Update the SVN and you should be good to go! Remember to ty my SWEPS via SVN or download on gmod.org! They do not require GDC SVN to work.